requirements.txt
# ========= mllm-hallucination: requirements =========
# Notes:
# - PyTorch build varies by CUDA. We leave torch/vision/torchaudio commented.
#   Install them per your CUDA from https://pytorch.org/get-started/locally/ .
# - On CPU-only systems, one can use 'faiss-cpu'; on GPU systems, switch to 'faiss-gpu'.

# Core scientific stack
numpy>=2.0,<2.3
scipy>=1.11
pandas>=2.2
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
pyyaml>=6.0.1

# Plotting / media (for previews & GIF export)
imageio>=2.34
imageio-ffmpeg>=0.5
plotly>=5.20
kaleido==0.2.1           # for plotly static image export (optional; may need system deps)

# Data / HF ecosystem
datasets>=2.19
transformers>=4.43
accelerate>=0.33
huggingface-hub>=0.24
sentencepiece>=0.1.99
timm>=0.9.12
einops>=0.7.0

# Vision / audio models & tools used in pipelines
open-clip-torch>=2.24.0
openai-whisper>=20231117
librosa>=0.10
soundfile>=0.12.1

# KNN / graph / utils
faiss-cpu>=1.7.4         # switch to faiss-gpu on CUDA boxes
networkx>=3.2

# Optional: Jupyter niceties
ipykernel>=6.29

# ---------- Install PyTorch separately per CUDA ----------
# torch==2.3.1+cu121; sys_platform == 'linux'    # EXAMPLE ONLY (commented on purpose)
# torchvision==0.18.1+cu121; sys_platform == 'linux'
# torchaudio==2.3.1+cu121; sys_platform == 'linux'
# --------------------------------------------------------
